pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "logo-seg"
version = "0.1.0"
description = "Professional YOLO11 Logo Segmentation Toolkit"
readme = "README.md"
authors = [
    { name = "quangkmhd", email = "your.email@example.com" }
]
license = { text = "MIT" }
requires-python = ">=3.8"
dependencies = [
    "ultralytics",
    "gradio",
    "opencv-python",
    "python-dotenv",
    "pyyaml"
]

[project.scripts]
logo-seg = "main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.yaml", "*.pt"]
